Adapt configuration to use el-get as much as possible
Old elpa downloaded packages are removed and replaced with the el-get equivalents. All the elpa-to-submit packages are also removed, some of them are replaced with an equivalent. The loaddefs are no longer necessary. Some vendor packages are also moved to it. Instructions for the new package manager are documented in the README.